+++ CHANGES.rst
@@ -11,6 +11,7 @@ Unreleased
 -   Fixed the issue where typing requires template global
     decorators to accept functions with no arguments. :issue:`4098`
 -   Support View and MethodView instances with async handlers. :issue:`4112`
+-   Enhance typing of ``app.errorhandler`` decorator. :issue:`4095`
 
 
 Version 2.0.1